Upgrade to match BusyBox xworld/20240105/1_36_1
There is a severe problem with rpc.h and the libraries associated with
it.

Linking them to BusyBox would create a lot of shared library
dependencies, and the only reason why rpc.h would be required ist RPC
support in the inetd applet.

I tried to link those libraries statically in order to avoid the shared
library dependencies, but this did not work because the developer
package did not seem to provide static archives for all shared libraries
included.

Therefore do the only sane thing right now and disable RPC-support for
inetd.

I will re-enable it once I find a way how to obtain the missing static
libraries.